Rename package and references to ebmbot.db
