+++ bennettbot/connection.py
