+++ bennettbot/job_configs.py
@@ -5,7 +5,7 @@ Defines which jobs may be run, and how they can be invoked via Slack.
 import re
 from operator import itemgetter
 
-from ebmbot import settings
+from bennettbot import settings
 
 
 # fmt: off
