+++ bennettbot/signatures.py
