+++ bennettbot/slack.py
@@ -2,7 +2,7 @@ from time import sleep
 
 from slack_sdk import WebClient
 
-from ebmbot import settings
+from bennettbot import settings
 from workspace.utils.blocks import get_basic_header_and_text_blocks, truncate_text
 
 from .logger import logger
