+++ bennettbot/webserver/__main__.py
@@ -6,6 +6,6 @@ from . import app
 
 
 if __name__ == "__main__":
-    logger.info("running ebmbot.webserver")
+    logger.info("running bennettbot.webserver")
     port = urlparse(settings.WEBHOOK_ORIGIN).port
     app.run(host="0.0.0.0", port=port, load_dotenv=False, debug=False)
