+++ workspace/workflows/jobs.py
@@ -6,7 +6,7 @@ from urllib.parse import urljoin
 
 import requests
 
-from ebmbot import settings
+from bennettbot import settings
 from workspace.utils.blocks import (
     get_basic_header_and_text_blocks,
     get_header_block,
